Update bootloader and x86_64 packages
This at least gets the kernel to the point where we can print to
a console. Hardware interrupts still need to be fixed by swapping
out the legacy PICS implementation with an APIC implementation.
Dynamic memory allocation also needs to be fixed.

The new workspace organization also prevents us from running tests,
so those need to be fixed too.